# docs/openapi.json
{
  "components": {
    "schemas": {
      "Assignment": {
        "properties": {
          "child_id": {
            "format": "uuid",
            "title": "Child Id",
            "type": "string"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "default": "practice",
            "title": "Mode",
            "type": "string"
          },
          "question_set_id": {
            "format": "uuid",
            "title": "Question Set Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AssignmentStatus",
            "default": "assigned"
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds"
          }
        },
        "required": [
          "family_id",
          "question_set_id",
          "child_id"
        ],
        "title": "Assignment",
        "type": "object"
      },
      "AssignmentStatus": {
        "enum": [
          "assigned",
          "in_progress",
          "submitted",
          "grading",
          "results_ready",
          "correcting",
          "completed",
          "withdrawn",
          "stopped"
        ],
        "title": "AssignmentStatus",
        "type": "string"
      },
...
      "Child": {
        "properties": {
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "grade_stage": {
            "title": "Grade Stage",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "nickname": {
            "title": "Nickname",
            "type": "string"
          },
          "ui_language": {
            "default": "en",
            "title": "Ui Language",
            "type": "string"
          }
        },
        "required": [
          "family_id",
          "nickname",
          "grade_stage"
        ],
        "title": "Child",
        "type": "object"
      },
...
      "ChildSessionRequest": {
        "properties": {
          "pin": {
            "pattern": "^\\d{6}$",
            "title": "Pin",
            "type": "string"
          }
        },
        "required": [
          "pin"
        ],
        "title": "ChildSessionRequest",
        "type": "object"
      },
      "ChildSessionResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "child_id": {
            "format": "uuid",
            "title": "Child Id",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "nickname": {
            "title": "Nickname",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "ui_language": {
            "enum": [
              "zh",
              "ja",
              "en"
            ],
            "title": "Ui Language",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "child_id",
          "family_id",
          "nickname",
          "ui_language"
        ],
        "title": "ChildSessionResponse",
        "type": "object"
      },
...
      "DemoBootstrap": {
        "properties": {
          "assignment": {
            "$ref": "#/components/schemas/Assignment"
          },
          "child": {
            "$ref": "#/components/schemas/Child"
          },
          "family": {
            "$ref": "#/components/schemas/Family"
          },
          "question_set": {
            "$ref": "#/components/schemas/QuestionSet"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionView"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "family",
          "child",
          "question_set",
          "assignment",
          "questions"
        ],
        "title": "DemoBootstrap",
        "type": "object"
      },
      "Family": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "Family",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Job": {
        "properties": {
          "attempt_count": {
            "default": 0,
            "title": "Attempt Count",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "queued"
          },
          "subject_id": {
            "format": "uuid",
            "title": "Subject Id",
            "type": "string"
          },
          "type": {
            "default": "grade_submission",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "family_id",
          "subject_id"
        ],
        "title": "Job",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "QuestionSet": {
        "properties": {
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionSetStatus"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "family_id",
          "title",
          "subject",
          "status"
        ],
        "title": "QuestionSet",
        "type": "object"
      },
...
      "QuestionSetStatus": {
        "enum": [
          "draft",
          "processing",
          "needs_review",
          "confirmed"
        ],
        "title": "QuestionSetStatus",
        "type": "string"
      },
      "QuestionType": {
        "enum": [
          "single_choice",
          "multiple_choice",
          "typed_text",
          "word_order",
          "handwriting",
          "photo",
          "listening"
        ],
        "title": "QuestionType",
        "type": "string"
      },
      "QuestionView": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "points": {
            "title": "Points",
            "type": "number"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/QuestionType"
          }
        },
        "required": [
          "id",
          "position",
          "type",
          "prompt",
          "points"
        ],
        "title": "QuestionView",
        "type": "object"
      },
...
      "UpdateChildPinRequest": {
        "properties": {
          "pin": {
            "pattern": "^\\d{6}$",
            "title": "Pin",
            "type": "string"
          }
        },
        "required": [
          "pin"
        ],
        "title": "UpdateChildPinRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Learning Assessment API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/children/{child_id}/pin": {
      "put": {
        "operationId": "update_child_pin_v1_children__child_id__pin_put",
        "parameters": [
          {
            "in": "path",
            "name": "child_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Child Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Management-Unlock",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Management-Unlock"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChildPinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Child Pin",
        "tags": [
          "children"
        ]
      }
    },
    "/v1/children/{child_id}/sessions": {
      "post": {
        "operationId": "create_child_session_v1_children__child_id__sessions_post",
        "parameters": [
          {
            "in": "path",
            "name": "child_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Child Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Child Session",
        "tags": [
          "children"
        ]
      }
    },
...
    "/v1/demo/bootstrap": {
      "post": {
        "operationId": "bootstrap_demo_v1_demo_bootstrap_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoBootstrap"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bootstrap Demo",
        "tags": [
          "demo"
        ]
      }
    },
    "/v1/demo/jobs/process-next": {
      "post": {
        "operationId": "process_next_fixture_job_v1_demo_jobs_process_next_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Process Next Fixture Job",
        "tags": [
          "demo"
        ]
      }
    },
...
  }
}